guess logic and title moved to provider now
